@@ -4,10 +4,10 @@
   "private": true,
   "dependencies": {
     "node-sass": "^4.11.0",
-    "react": "^16.7.0",
+    "react": "^16.8.1",
     "react-app-rewire-less": "^2.1.3",
     "react-app-rewired": "^2.0.1",
-    "react-dom": "^16.7.0",
+    "react-dom": "^16.8.1",
     "react-scripts": "2.1.3"
   },
   "scripts": {
